Install xquartz via direct download rather than homebrew cask.
We want homebrew operations to run via the Jenkins user to keep
permissions in good order but the Chef executor cannot do so without a
password prompt which fails. So this is just easier.

The biggest downside is that it does not monitor/maintain updates but
neither does homebrew-cask by default either so :shrug:.