Remove default labels array and add docker label at runtime.
In order to allow re-use across different platforms, the docker label
shouldn't be the default. Since there is no default, let's leave it off.
This means we need to handle its absence in the recipe.

We also add the docker label (if it is not already present)
automatically for all hosts since all Linux hosts are docker agents
currently.