Include g5.xlarge instances in generic GPU detection.
The g5.xlarge instances we are trying to deploy have the following lspci
output:

```
00:00.0 Host bridge: Intel Corporation 440FX - 82441FX PMC [Natoma]
00:01.0 ISA bridge: Intel Corporation 82371SB PIIX3 ISA [Natoma/Triton II]
00:01.3 Non-VGA unclassified device: Intel Corporation 82371AB/EB/MB PIIX4 ACPI (rev 08)
00:03.0 VGA compatible controller: Amazon.com, Inc. Device 1111
00:04.0 Non-Volatile memory controller: Amazon.com, Inc. Device 8061
00:05.0 Ethernet controller: Amazon.com, Inc. Elastic Network Adapter (ENA)
00:1e.0 3D controller: NVIDIA Corporation Device 2237 (rev a1)
00:1f.0 Non-Volatile memory controller: Amazon.com, Inc. NVMe SSD Controller
 ```

(Note the use of 3D controller rather than VGA compatible controller)
This updates the nvidia detection to try and use this.

Since this does not match the nvidia grid output from the existing AWS
support I did not update that to use it.

Further iteration may be required to get these working.