Add a cookbook ohai plugin for doing GPU detection in ohai.
We have some helpers shelling out to lspci during the cookbook run,
however, as we extend the amount of information we're getting from lspci
we also need to use it in more places. Getting the information in ohai
means that we can use it via node attributes all around our chef recipes
in a consistent fashion.