Do not apply docker label to all agents (#41)
The docker label has traditionally implied in the buildfarm that the agent has docker on top of amd64 system. We had the problem of adding docker to arm agents making them to be used as debbuilders and generating arm packages instead of the amd64 expected ones. The PR removes the addition of docker by default to all agents, something that should be controlled via role.

Run CI on all branches

---------

Signed-off-by: Jose Luis Rivero <jrivero@osrfoundation.org>